Leveling With Base (#1)
* Added Udacity Lessons for Widgets & Espresso

These lessons are free and very nicely explained. So, I think it would be better to have them in the list.

* Added the Udacity lesson about "Integrate code from an external support library"

Just saw that this section was missing lessons (It was empty). So, I thought that I should add a great lesson about it in the list.

* Adding few missing "checkboxes"

The article/readme was looking little inconsistent with "bullets (•)" instead of "checkboxes" in some of the subtopics/subheadings. So, I thought that I should correct this.

* Minor Correction in Youtube Links

Corrected a typo (missing "left square bracket" in a link). Also, there were few inconsistency in "Youtube Links" naming. Just corrected them.

* Added "Fragments topic" links to the List

I appeared for the Associate Android Developer Certification just few weeks ago. I was asked questions about "Fragments" in the "Exit Interview" of the AAD Certification (I cannot share the exact questions here in public). But, what I just want to say is that "Fragments" topic is important.

* Update README.md

Just added some Contents and added materials including Room,LiveData,Lifecycle and Paging Library

* Correcting minor grammar issue

* Adding more references to Architecture Components

* Adding credits  to Pheonix73

* Added new sources (#11)

* Update README.md

* Update README.md (#12)

Fixed minor grammatical issue

* Fixed typos (#13)

Co-authored-by: Suryakant Bharti <suryakant.bharti@gmail.com>
Co-authored-by: Arturo Mejia <arturomejiamarmol@gmail.com>
Co-authored-by: Kswiss73 <36895007+Pheonix73@users.noreply.github.com>
Co-authored-by: psbankar <prasadbankar1997@gmail.com>
Co-authored-by: RajaVamsi11 <30550059+RajaVamsi11@users.noreply.github.com>
Co-authored-by: Rishabh Maurya <rishabhmaurya.maurya@gmail.com>